+++ docs/changelog.rst
@@ -14,6 +14,7 @@ CHANGELOG
 * Home now redirects to treks list in light version (without topologies)
 * Fix Cirkwi export in light version
 * Fix SRID in database migrations
+* Add signage type filter again (fixes #1352)
 
 
 0.32.1 (2015-03-04)